# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Authentication and Authorization API",
        "title": "Auth API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/admin/jobs": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Merge, split, compress va boshqa joblar holati (filter + paginatsiya)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin",
                    "jobs"
                ],
                "summary": "List jobs (all types)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job type (merge|split|compress)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Status (pending|processing|done|failed)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "From (RFC3339, e.g. 2025-08-01T00:00:00Z)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "To (RFC3339)",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Job ID prefix",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Limit (max 200)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.JobSummary"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/pdf/add-page-numbers": {
            "post": {
                "description": "Adds page numbers to selected pages with font/position settings (auth optional).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pdf"
                ],
                "summary": "Add page numbers to a PDF",
                "parameters": [
                    {
                        "description": "PDF ID and font settings",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddPageNumbersRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "job_id",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/pdf/add-page-numbers/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pdf"
                ],
                "summary": "Get add-page-numbers job by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AddPageNumberJob"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/pdf/compress": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pdf-compress"
                ],
                "summary": "Create compress job",
                "parameters": [
                    {
                        "description": "compress job",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CompressRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/pdf/compress/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pdf-compress"
                ],
                "summary": "Get compress job by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "compress job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CompressJob"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/pdf/watermark": {
            "post": {
                "description": "Add a watermark text to specific pages of a PDF file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pdf-watermark"
                ],
                "summary": "Add text watermark to PDF",
                "parameters": [
                    {
                        "description": "Watermark request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddWatermarkRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/pdf/watermark/{id}": {
            "get": {
                "description": "Retrieve watermark job details and status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pdf-watermark"
                ],
                "summary": "Get watermark job by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Watermark Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AddWatermarkJob"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.AddPageNumberJob": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "first_number": {
                    "type": "integer"
                },
                "font_size": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "input_file_id": {
                    "type": "string"
                },
                "output_file_id": {
                    "type": "string"
                },
                "page_range": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.AddPageNumbersRequest": {
            "type": "object",
            "required": [
                "color",
                "first_number",
                "font_size",
                "input_file_id",
                "page_range",
                "position"
            ],
            "properties": {
                "color": {
                    "type": "string"
                },
                "first_number": {
                    "description": "\u003c-- bu yer o‘zgardi",
                    "type": "integer",
                    "minimum": 1
                },
                "font_size": {
                    "description": "\u003c-- bu yer o‘zgardi",
                    "type": "integer",
                    "minimum": 1
                },
                "input_file_id": {
                    "type": "string"
                },
                "page_range": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                }
            }
        },
        "models.AddWatermarkJob": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "fill_color": {
                    "type": "string"
                },
                "font_name": {
                    "type": "string"
                },
                "font_size": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "input_file_id": {
                    "type": "string"
                },
                "opacity": {
                    "type": "number"
                },
                "output_file_id": {
                    "type": "string"
                },
                "pages": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                },
                "rotation": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.AddWatermarkRequest": {
            "type": "object",
            "required": [
                "fill_color",
                "font_name",
                "font_size",
                "input_file_id",
                "position",
                "text"
            ],
            "properties": {
                "fill_color": {
                    "description": "e.g., \"#FF0000\"",
                    "type": "string"
                },
                "font_name": {
                    "description": "e.g., Helvetica",
                    "type": "string"
                },
                "font_size": {
                    "description": "e.g., 24",
                    "type": "integer",
                    "maximum": 72,
                    "minimum": 8
                },
                "input_file_id": {
                    "type": "string"
                },
                "opacity": {
                    "description": "e.g., 0.6",
                    "type": "number",
                    "maximum": 1,
                    "minimum": 0
                },
                "pages": {
                    "description": "optional: \"1-3\"",
                    "type": "string"
                },
                "position": {
                    "description": "e.g., \"tl\"",
                    "type": "string"
                },
                "rotation": {
                    "description": "e.g., 0",
                    "type": "integer",
                    "maximum": 360,
                    "minimum": 0
                },
                "text": {
                    "type": "string"
                }
            }
        },
...
        "models.CompressJob": {
            "type": "object",
            "properties": {
                "compression": {
                    "description": "\"low\", \"medium\", \"high\"",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.CompressionLevel"
                        }
                    ]
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "input_file_id": {
                    "type": "string"
                },
                "output_file_id": {
                    "description": "Output file, can be nil",
                    "type": "string"
                },
                "status": {
                    "description": "\"pending\", \"processing\", \"done\", \"failed\"",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.JobStatus"
                        }
                    ]
                },
                "user_id": {
                    "description": "Can be nil for guest users",
                    "type": "string"
                }
            }
        },
        "models.CompressRequest": {
            "type": "object",
            "required": [
                "compression",
                "input_file_id"
            ],
            "properties": {
                "compression": {
                    "description": "\"low\", \"medium\", \"high\"",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.CompressionLevel"
                        }
                    ]
                },
                "input_file_id": {
                    "type": "string"
                }
            }
        },
        "models.CompressionLevel": {
            "type": "string",
            "enum": [
                "low",
                "medium",
                "high"
            ],
            "x-enum-varnames": [
                "Low",
                "Medium",
                "High"
            ]
        },
...
        "models.JobStatus": {
            "type": "string",
            "enum": [
                "pending",
                "processing",
                "done",
                "failed"
            ],
            "x-enum-varnames": [
                "Pending",
                "Processing",
                "Done",
                "Failed"
            ]
        },
        "models.JobSummary": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "job_type": {
                    "type": "string"
                },
                "output_file_id": {
                    "description": "split uchun NULL bo'lishi mumkin",
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "models.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "description": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}